refactor: extract handler functions
